Create new file
The idea will be to provide a build service via travis for easy testing of PRs.

Currently, it is necessary to download or git clone the repo in order to generate a customized x3dom.js. x3dom uses python to build and minimize x3dom.js. The provided python script has to be run from the command line.

PRs against this branch will be picked up automatically by Travis which is then instructed to git clone the branch, apply the PR, run the build python script, and upload the resulting x3dom.js and co. files to x3dom-builder.surge.sh.

By the default travis does not do the deploy phase for PRs. The solution will be to figure how to deploy from the test phase.