sync with x3dom/x3dom/master (#20)
* Let <code> elements wrap white space

Earlier today I tried reading your documentation page for IndexedFaceSet on my iPad. The entire page was zoomed out to accommodate the <code> element near the top, making it impossible to read without zooming in. Even for browsers that do not zoom out for extremely long lines, most of the text will be outside of the visible window and not easily accessible.

This small change keeps the HTML encoding of the node and all of its defaults in a compact wrapped area where everything is easily seen, making the documentation more useful.

* fixesuggested by  LGTM (#747)

* declare local variables in three files
* cleaning

* check correctly for existance of boolean fields

a boolean field with a default value of 'false' was not recognized as existing; use explicit check for undefined

* formatting

* touchTime (#784)

* sync 
* add touchTime field event